finish? datafix.do
goals for data fix as I understand them have been accomplished as far as they reasonably can be. I made the determination that of the relevant data we're using are either in good condition or have been to as close as usable as they can. The fix on annual average employees I've used could be refined and expanded but other than that we're good